Rename associated to associated_scope_validator
