+++ lib/associated_scope_validator.rb
